moved the input.txt to resources, and am locating the file using URL getResource. it finds it in the target/class folder and counts and modifies that txt file to show the unique count.
