+++ src/main/java/com/solvd/airport/CarryOnBag.java
@@ -2,10 +2,7 @@ package com.solvd.airport;
 
 public class CarryOnBag extends Luggage {
 
-    private final String LUGGAGE_TYPE = "CarryOn";
-
     public CarryOnBag(int bagCount, int weight) {
         super(bagCount, weight);
     }
-
 }
