+++ src/main/java/com/solvd/airport/FoodCourt.java
@@ -12,7 +12,6 @@ import java.util.List;
 public class FoodCourt implements IFacility {
 
     private static final Logger LOGGER = LogManager.getLogger(MethodHandles.lookup().lookupClass());
-
     private String foodType;
     private String restaurantName;
     private double price;
