--- src/main/java/com/solvd/exceptions/OverWeightBagException.java
@@ -1,7 +1,0 @@
-package com.solvd.exceptions;
-
-public class OverWeightBagException extends Exception {
-    public OverWeightBagException(String message) {
-        super(message);
-    }
-}
