+++ src/main/java/com/solvd/utils/FileProcessor.java
@@ -5,7 +5,8 @@ import org.apache.commons.lang3.StringUtils;
 
 import java.io.File;
 import java.io.IOException;
-
+import java.net.URL;
+import java.net.URISyntaxException;
 import java.util.Set;
 import java.util.HashSet;
 
@@ -15,16 +16,23 @@ import java.util.List;
 
 public class FileProcessor {
 
-    public static void main(String[] args) {
+    public static void main(String[] args) throws URISyntaxException {
         //define the input file
-        File inputFile = new File("src/main/java/com/solvd/utils/input.txt");
+        URL resource = FileProcessor.class.getClassLoader().getResource("input.txt");
+        if (resource == null) {
+            System.err.println("File not found!");
+            return;
+        }
+
+        File inputFile = new File(resource.getFile());
+        System.out.println(inputFile);
         try {
             //read all lines from the input file
             List<String> lines = FileUtils.readLines(inputFile, "UTF-8");
 
             //process the lines to get all words
-            String text = String.join(" ", lines);
-            String[] words = StringUtils.split(text, ' ');
+            String text = String.join(" ", lines); //join lines into a single text
+            String[] words = StringUtils.split(text, ' '); //split text into an array of words
 
             //calc number of unique words
             int uniqueWordCount = countUniqueWords(words);
@@ -32,7 +40,8 @@ public class FileProcessor {
             //Write the result to the same file
             FileUtils.write(inputFile, "\nnumber of unique words: " + uniqueWordCount, "UTF-8", true);
         } catch (IOException e) {
-            e.printStackTrace();
+            //handle IO exception
+            e.printStackTrace();//print the exception details to the console
         }
     }
