+++ src/main/resources/input.txt
@@ -1,0 +1,1 @@
+hello world file
